Fix some issues with moving around token prices
