+++ edsl/jobs/interviews/InterviewStatistic.py
@@ -5,7 +5,7 @@ from collections import UserDict, defaultdict
 
 from edsl.jobs.interviews.InterviewStatusDictionary import InterviewStatusDictionary
 from edsl.jobs.tokens.InterviewTokenUsage import InterviewTokenUsage
-from edsl.jobs.pricing import pricing, TokenPricing
+from edsl.enums import pricing, TokenPricing
 from edsl.jobs.tasks.task_status_enum import TaskStatus
 
 InterviewTokenUsageMapping = DefaultDict[str, InterviewTokenUsage]
